# docs/OPENAPI.JSON
{
  "openapi": "3.1.0",
  "info": {
    "title": "Savory Canvas API",
    "version": "1.2.1",
    "description": "Savory Canvas 本地图文生成系统接口文档"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8887",
      "description": "本地开发环境"
    }
  ],
...
  "paths": {
...
    "/api/v1/inspirations/{session_id}": {
      "get": {
        "tags": [
          "Inspiration"
        ],
        "summary": "获取灵感对话上下文",
        "operationId": "getInspirationConversation",
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspirationConversationResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/inspirations/messages": {
      "post": {
        "tags": [
          "Inspiration"
        ],
        "summary": "发送灵感对话消息（支持文本/图片/视频混合）",
        "operationId": "postInspirationMessage",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "session_id"
                ],
                "properties": {
                  "session_id": {
                    "type": "string"
                  },
                  "text": {
                    "type": "string"
                  },
                  "selected_items": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "action": {
                    "type": "string",
                    "enum": [
                      "continue",
                      "confirm_prompt",
                      "confirm_assets",
                      "revise_assets",
                      "use_style_profile",
                      "save_style",
                      "skip_save"
                    ]
                  },
                  "image_usages": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "style_reference",
                        "content_asset"
                      ]
                    },
                    "description": "与 images 按顺序对应的图片用途标签。未提供时默认 content_asset。"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  },
                  "videos": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "处理成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspirationConversationResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/assets/text": {
      "post": {
        "tags": [
          "Asset"
        ],
        "summary": "提交文本素材",
        "operationId": "createTextAsset",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetTextCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "提交成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/assets/video": {
      "post": {
        "tags": [
          "Asset"
        ],
        "summary": "上传视频素材",
        "operationId": "uploadVideoAsset",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "session_id",
                  "file"
                ],
                "properties": {
                  "session_id": {
                    "type": "string"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "上传成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/assets/{asset_id}/transcript": {
      "get": {
        "tags": [
          "Asset"
        ],
        "summary": "转写结果",
        "operationId": "getTranscript",
        "parameters": [
          {
            "$ref": "#/components/parameters/AssetId"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptResult"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
    "/api/v1/jobs/generate": {
      "post": {
        "tags": [
          "Generation"
        ],
        "summary": "创建生成任务",
        "operationId": "createGenerationJob",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerationJobCreateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "任务已创建",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationJob"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "Generation"
        ],
        "summary": "任务状态",
        "operationId": "getGenerationJob",
        "parameters": [
          {
            "$ref": "#/components/parameters/JobId"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationJob"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/stages": {
      "get": {
        "tags": [
          "Generation"
        ],
        "summary": "任务阶段轨迹",
        "operationId": "getGenerationJobStages",
        "parameters": [
          {
            "$ref": "#/components/parameters/JobId"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationStageListResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/asset-breakdown": {
      "get": {
        "tags": [
          "Generation"
        ],
        "summary": "任务资产拆解结果",
        "operationId": "getGenerationJobAssetBreakdown",
        "parameters": [
          {
            "$ref": "#/components/parameters/JobId"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAssetBreakdownResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/results": {
      "get": {
        "tags": [
          "Generation"
        ],
        "summary": "任务结果",
        "operationId": "getGenerationResult",
        "parameters": [
          {
            "$ref": "#/components/parameters/JobId"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationResult"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "Generation"
        ],
        "summary": "取消生成任务",
        "operationId": "cancelGenerationJob",
        "parameters": [
          {
            "$ref": "#/components/parameters/JobId"
          }
        ],
        "responses": {
          "202": {
            "description": "已提交取消请求",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationJob"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/exports": {
      "post": {
        "tags": [
          "Export"
        ],
        "summary": "创建导出任务",
        "operationId": "createExportTask",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportTaskCreateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "导出任务已创建",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportTask"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/exports/{export_id}": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "导出任务状态",
        "operationId": "getExportTask",
        "parameters": [
          {
            "$ref": "#/components/parameters/ExportId"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportTask"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
    "/api/v1/models": {
      "get": {
        "tags": [
          "Model"
        ],
        "summary": "模型列表",
        "operationId": "listModels",
        "parameters": [
          {
            "name": "provider_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "提供商 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelListResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/config/model-routing": {
      "get": {
        "tags": [
          "Model"
        ],
        "summary": "获取当前模型路由",
        "operationId": "getModelRouting",
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ModelRoutingConfig"
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      },
      "post": {
        "tags": [
          "Model"
        ],
        "summary": "设置模型路由",
        "operationId": "updateModelRouting",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelRoutingUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "设置成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelRoutingConfig"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "SessionId": {
        "name": "session_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "AssetId": {
        "name": "asset_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
      "JobId": {
        "name": "job_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "ExportId": {
        "name": "export_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
    "responses": {
      "ErrorResponse": {
        "description": "错误响应",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "code",
                "message"
              ],
              "properties": {
                "code": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                },
                "details": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "AssetTextCreateRequest": {
        "type": "object",
        "required": [
          "session_id",
          "asset_type",
          "content"
        ],
        "properties": {
          "session_id": {
            "type": "string"
          },
          "asset_type": {
            "type": "string",
            "enum": [
              "food_name",
              "scenic_name",
              "text"
            ]
          },
          "content": {
            "type": "string"
          }
        }
      },
      "Asset": {
        "type": "object",
        "required": [
          "id",
          "session_id",
          "asset_type",
          "status",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          },
          "asset_type": {
            "type": "string",
            "enum": [
              "food_name",
              "scenic_name",
              "text",
              "image",
              "video",
              "transcript"
            ]
          },
          "content": {
            "type": "string"
          },
          "file_path": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "processing",
              "failed"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "InspirationMessage": {
        "type": "object",
        "required": [
          "id",
          "role",
          "content",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "assistant",
              "user",
              "system"
            ]
          },
          "content": {
            "type": "string"
          },
          "options": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StyleOptionBlock"
              },
              {
                "type": "null"
              }
            ]
          },
          "fallback_used": {
            "type": "boolean"
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InspirationAttachment"
            }
          },
          "asset_candidates": {
            "$ref": "#/components/schemas/InspirationAssetCandidates"
          },
          "style_context": {
            "$ref": "#/components/schemas/InspirationStyleContext"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "InspirationDraft": {
        "type": "object",
        "required": [
          "stage",
          "locked"
        ],
        "properties": {
          "stage": {
            "type": "string",
            "enum": [
              "style_collecting",
              "prompt_revision",
              "asset_confirming",
              "locked"
            ]
          },
          "style_payload": {
            "$ref": "#/components/schemas/StylePayload"
          },
          "image_count": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10
          },
          "draft_style_id": {
            "type": "string"
          },
          "locked": {
            "type": "boolean"
          }
        }
      },
      "InspirationConversationResponse": {
        "type": "object",
        "required": [
          "session_id",
          "messages",
          "draft"
        ],
        "properties": {
          "session_id": {
            "type": "string"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InspirationMessage"
            }
          },
          "draft": {
            "$ref": "#/components/schemas/InspirationDraft"
          }
        }
      },
      "TranscriptResult": {
        "type": "object",
        "required": [
          "asset_id",
          "status"
        ],
        "properties": {
          "asset_id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "processing",
              "ready",
              "failed"
            ]
          },
          "text": {
            "type": "string"
          },
          "segments": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
...
      "GenerationJobCreateRequest": {
        "type": "object",
        "required": [
          "session_id",
          "style_profile_id",
          "image_count"
        ],
        "properties": {
          "session_id": {
            "type": "string"
          },
          "style_profile_id": {
            "type": "string"
          },
          "image_count": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10
          }
        }
      },
      "GenerationJob": {
        "type": "object",
        "required": [
          "id",
          "session_id",
          "status",
          "progress_percent",
          "current_stage",
          "stage_message",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          },
          "style_profile_id": {
            "type": "string"
          },
          "image_count": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "partial_success",
              "success",
              "failed",
              "canceled"
            ]
          },
          "progress_percent": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "current_stage": {
            "type": "string",
            "enum": [
              "asset_extract",
              "asset_allocate",
              "prompt_generate",
              "image_generate",
              "copy_generate",
              "finalize"
            ]
          },
          "stage_message": {
            "type": "string"
          },
          "error_code": {
            "type": "string"
          },
          "error_message": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "JobAssetBreakdownResponse": {
        "type": "object",
        "required": [
          "job_id",
          "session_id",
          "content_mode",
          "source_assets",
          "extracted",
          "created_at"
        ],
        "properties": {
          "job_id": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          },
          "content_mode": {
            "type": "string",
            "enum": [
              "food",
              "scenic",
              "food_scenic"
            ]
          },
          "source_assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssetBreakdownSourceAsset"
            }
          },
          "extracted": {
            "$ref": "#/components/schemas/AssetBreakdownExtracted"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "GenerationResult": {
        "type": "object",
        "required": [
          "job_id",
          "status",
          "images",
          "copy"
        ],
        "properties": {
          "job_id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImageResult"
            }
          },
          "copy": {
            "$ref": "#/components/schemas/CopyResult"
          }
        }
      },
      "ExportTaskCreateRequest": {
        "type": "object",
        "required": [
          "session_id",
          "job_id",
          "export_format"
        ],
        "properties": {
          "session_id": {
            "type": "string"
          },
          "job_id": {
            "type": "string"
          },
          "export_format": {
            "type": "string",
            "enum": [
              "long_image",
              "pdf"
            ]
          }
        }
      },
      "ExportTask": {
        "type": "object",
        "required": [
          "id",
          "session_id",
          "job_id",
          "export_format",
          "status",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          },
          "job_id": {
            "type": "string"
          },
          "export_format": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "success",
              "failed"
            ]
          },
          "file_url": {
            "type": "string"
          },
          "error_code": {
            "type": "string"
          },
          "error_message": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ModelListResponse": {
        "type": "object",
        "required": [
          "provider_id",
          "items"
        ],
        "properties": {
          "provider_id": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelInfo"
            }
          }
        }
      },
...
      "ModelRoutingUpdateRequest": {
        "type": "object",
        "required": [
          "image_model",
          "text_model"
        ],
        "properties": {
          "image_model": {
            "$ref": "#/components/schemas/ModelReference"
          },
          "text_model": {
            "$ref": "#/components/schemas/ModelReference"
          }
        }
      },
      "ModelRoutingConfig": {
        "type": "object",
        "required": [
          "image_model",
          "text_model",
          "updated_at"
        ],
        "properties": {
          "image_model": {
            "$ref": "#/components/schemas/ModelReference"
          },
          "text_model": {
            "$ref": "#/components/schemas/ModelReference"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "GenerationStageItem": {
        "type": "object",
        "required": [
          "stage",
          "status",
          "stage_message",
          "created_at"
        ],
        "properties": {
          "stage": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "stage_message": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "GenerationStageListResponse": {
        "type": "object",
        "required": [
          "job_id",
          "items"
        ],
        "properties": {
          "job_id": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenerationStageItem"
            }
          }
        }
      },
      "AssetBreakdownSourceAsset": {
        "type": "object",
        "required": [
          "asset_id",
          "asset_type"
        ],
        "properties": {
          "asset_id": {
            "type": "string"
          },
          "asset_type": {
            "type": "string"
          },
          "content": {
            "type": "string"
          }
        }
      },
...
    }
  }
}